// frontend/plugins/welcome/src/api/models/EntPlace.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntPlaceEdges,
    EntPlaceEdgesFromJSON,
    EntPlaceEdgesFromJSONTyped,
    EntPlaceEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntPlace
 */
export interface EntPlace {
    /**
     * PlaceName holds the value of the "PlaceName" field.
     * @type {string}
     * @memberof EntPlace
     */
    placeName?: string;
    /**
     * 
     * @type {EntPlaceEdges}
     * @memberof EntPlace
     */
    edges?: EntPlaceEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntPlace
     */
    id?: number;
}

export function EntPlaceFromJSON(json: any): EntPlace {
    return EntPlaceFromJSONTyped(json, false);
}

export function EntPlaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntPlace {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'placeName': !exists(json, 'PlaceName') ? undefined : json['PlaceName'],
        'edges': !exists(json, 'edges') ? undefined : EntPlaceEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntPlaceToJSON(value?: EntPlace | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'PlaceName': value.placeName,
        'edges': EntPlaceEdgesToJSON(value.edges),
        'id': value.id,
    };
}